로그아웃 기능 추가

+++ src/main/resources/static/js/profile.js
@@ -19,24 +19,23 @@ let callbackToUpdate = (ev)=> {
 if(resumeAddButton != null) {
     resumeAddButton.addEventListener("click", callbackToUpdate);
 }
-if(resumeUpdateButton != null) {
+
+if(document.getElementById("resumeBox")) {
+    document.getElementById("resumeBox").addEventListener("click", (ev) => {
+        location = `/resume?email=${ownerEmail}`;
+    });
     resumeUpdateButton.addEventListener("click", (ev) => {
         alert("이력서를 수정합니다.");
         location=`/resume-update?email=${ownerEmail}`;
     });
-};
-
-document.getElementById("resumeBox").addEventListener("click", (ev) => {
-    location = `/resume?email=${ownerEmail}`;
-});
-
-document.getElementById("resumeDeleteButton").addEventListener("click", (ev) => {
-    ev.stopPropagation();
-    const userResponse = confirm('정말로 이력서를 삭제하시겠습니까?');
-    if(!userResponse) return;
-    fetch(`/api/resume?email=${ownerEmail}`, {
-        method :'DELETE'
-    }).then((ev) => {
-        location = `/profile?email=${ownerEmail}`;
+    document.getElementById("resumeDeleteButton").addEventListener("click", (ev) => {
+        ev.stopPropagation();
+        const userResponse = confirm('정말로 이력서를 삭제하시겠습니까?');
+        if(!userResponse) return;
+        fetch(`/api/resume?email=${ownerEmail}`, {
+            method :'DELETE'
+        }).then((ev) => {
+            location = `/profile?email=${ownerEmail}`;
+        });
     });
-});
+}
